# backend/openapi_server/openapi/openapi.yaml
openapi: 3.0.3
info:
  description: |
    api spec for StoryNet, covering the backend API and the internal apis of the pdf_service and ner_service. all public endpoints require JWT authentication unless explicitly marked otherwise.
  title: StoryNet API
  version: 2.0.0
servers:
#- description: backend API
#  url: https://angles-server.ddns.net:5001/api
- description: Internal pdf_service (OCR) — not publicly reachable
  url: http://pdf-service.app.svc.cluster.local:5001
- description: Internal ner_service (NER/LLM) — not publicly reachable
  url: http://ner-service.app.svc.cluster.local:5000
security:
- bearerAuth: []
...
paths:
  /auth/login:
    post:
      operationId: auth_login_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
          description: Authentication successful.
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Invalid credentials.
      security: []
      summary: Authenticate and receive a JWT
      tags:
      - Auth
      x-openapi-router-controller: openapi_server.controllers.auth_controller
  /auth/me:
    get:
      operationId: auth_me_get
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
          description: Current user profile.
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Missing or invalid JWT.
      summary: Get the authenticated user's profile
      tags:
      - Auth
      x-openapi-router-controller: openapi_server.controllers.auth_controller
  /auth/me/password:
    put:
      operationId: auth_me_password_put
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/_auth_me_password_put_request"
        required: true
      responses:
        "204":
          description: Password changed.
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Missing or invalid JWT.
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Validation error.
      summary: Change the authenticated user's password
      tags:
      - Auth
      x-openapi-router-controller: openapi_server.controllers.auth_controller
  /auth/refresh:
    post:
      operationId: auth_refresh_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/_auth_refresh_post_request"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
          description: New token pair issued.
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Refresh token invalid or expired.
      security: []
      summary: Obtain a new access token using a refresh token
      tags:
      - Auth
      x-openapi-router-controller: openapi_server.controllers.auth_controller
  /auth/register:
    post:
      operationId: auth_register_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
          description: Account created. Returns JWT immediately.
        "409":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Email already registered.
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Validation error.
      security: []
      summary: Register a new user account
      tags:
      - Auth
      x-openapi-router-controller: openapi_server.controllers.auth_controller
...
components:
...
  schemas:
...
    LoginRequest:
      example:
        password: password
        email: email
      properties:
        email:
          format: email
          title: email
          type: string
        password:
          format: password
          title: password
          type: string
      required:
      - email
      - password
      title: LoginRequest
      type: object
    RegisterRequest:
      example:
        password: password
        display_name: display_name
        email: email
      properties:
        email:
          format: email
          title: email
          type: string
        password:
          format: password
          minLength: 8
          title: password
          type: string
        display_name:
          title: display_name
          type: string
      required:
      - display_name
      - email
      - password
      title: RegisterRequest
      type: object
    TokenResponse:
      example:
        access_token: access_token
        refresh_token: refresh_token
        token_type: Bearer
        expires_in: 0
      properties:
        access_token:
          description: JWT access token.
          title: access_token
          type: string
        refresh_token:
          title: refresh_token
          type: string
        token_type:
          enum:
          - Bearer
          title: token_type
          type: string
        expires_in:
          description: Seconds until access_token expires.
          title: expires_in
          type: integer
      required:
      - access_token
      - expires_in
      - token_type
      title: TokenResponse
      type: object
    UserProfile:
      example:
        created_at: 2000-01-23T04:56:07.000+00:00
        id: id
        display_name: display_name
        email: email
      properties:
        id:
          title: id
          type: string
        email:
          format: email
          title: email
          type: string
        display_name:
          title: display_name
          type: string
        created_at:
          format: date-time
          title: created_at
          type: string
      title: UserProfile
      type: object
...
    ErrorResponse:
      example:
        details:
          key: ""
        error: ENTITY_NOT_FOUND
        message: message
      properties:
        error:
          description: Machine-readable error code.
          example: ENTITY_NOT_FOUND
          title: error
          type: string
        message:
          description: Human-readable description.
          title: message
          type: string
        details:
          additionalProperties: true
          nullable: true
          title: details
          type: object
      required:
      - error
      - message
      title: ErrorResponse
      type: object
    _auth_refresh_post_request:
      properties:
        refresh_token:
          title: refresh_token
          type: string
      required:
      - refresh_token
      title: _auth_refresh_post_request
      type: object
    _auth_me_password_put_request:
      properties:
        current_password:
          format: password
          title: current_password
          type: string
        new_password:
          format: password
          minLength: 8
          title: new_password
          type: string
      required:
      - current_password
      - new_password
      title: _auth_me_password_put_request
      type: object
...
  securitySchemes:
    bearerAuth:
      bearerFormat: JWT
      scheme: bearer
      type: http
      x-bearerInfoFunc: openapi_server.controllers.security_controller.info_from_bearerAuth